requirements.txt
markdown>=3.4.1
python-docx>=0.8.11
reportlab>=4.0.4
Pillow>=9.0.0
